Update CommonStyles file:
- Add new 1200px maxWidth container.
- Extract HighlightSecondary component from welcome/Styles.js to
  reuse it.
